Utilise setup.cfg, move version info to redbot package
- `redbot.__init__` now is safe to import without installing dependencies.
- Now deploying binary wheel distribution from travis
