+++ setup.cfg
@@ -1,3 +1,103 @@
 [metadata]
+name = Red-DiscordBot
+version = attr: redbot.__version__
+description = A highly customisable Discord bot
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8; variant=GFM
+author = Cog-Creators
+author_email = cogcreators@gmail.com
+url = https://github.com/Cog-Creators/Red-DiscordBot
+classifiers =
+    Development Status :: 4 - Beta
+    Framework :: AsyncIO
+    Framework :: Pytest
+    Intended Audience :: Developers
+    License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3.7
+    Topic :: Communications :: Chat
+    Topic :: Documentation :: Sphinx
+
+[options]
+packages = find_namespace:
+python_requires = >=3.7,<3.8
+install_requires =
+    aiohttp-json-rpc==0.11.2
+    aiohttp==3.4.4
+    appdirs==1.4.3
+    async-timeout==3.0.1
+    attrs==18.2.0
+    chardet==3.0.4
+    colorama==0.4.1
+    distro==1.3.0; sys_platform == "linux"
+    fuzzywuzzy==0.17.0
+    idna-ssl==1.1.0
+    idna==2.8
+    multidict==4.5.2
+    python-levenshtein-wheels==0.13.1
+    pyyaml==3.13
+    raven==6.10.0
+    raven-aiohttp==0.7.0
+    red-lavalink==0.2.0
+    schema==0.6.8
+    websockets==6.0
+    yarl==1.3.0
+
+[options.extras_require]
+docs =
+    alabaster==0.7.12
+    babel==2.6.0
+    certifi==2018.11.29
+    docutils==0.14
+    imagesize==1.1.0
+    Jinja2==2.10
+    MarkupSafe==1.1.0
+    packaging==18.0
+    pyparsing==2.3.0
+    Pygments==2.3.1
+    pytz==2018.9
+    requests==2.21.0
+    six==1.12.0
+    snowballstemmer==1.2.1
+    sphinx==1.8.3
+    sphinx_rtd_theme==0.4.2
+    sphinxcontrib-asyncio==0.2.0
+    sphinxcontrib-websupport==1.1.0
+    urllib3==1.24.1
+mongo =
+    motor==2.0.0
+    pymongo==3.7.2
+    dnspython==1.16.0
+style =
+    black==18.9b0
+    click==7.0
+    toml==0.10.0
+test =
+    atomicwrites==1.2.1
+    more-itertools==5.0.0
+    pluggy==0.8.1
+    py==1.7.0
+    pytest==4.1.0
+    pytest-asyncio==0.10.0
+    six==1.12.0
+
+[options.entry_points]
+console_scripts =
+    redbot=redbot.__main__:main
+    redbot-setup=redbot.setup:main
+    redbot-launcher=redbot.launcher:main
+pytest11 =
+    red-discordbot=redbot.pytest
+
+[options.packages.find]
+include =
+    redbot
+    redbot.*
+    discord
+    discord.ext.commands
+
+[options.package_data]
+* =
+    locales/*.po
+    data/*
+    data/**/*
